% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bios524hwk6-package.R
\docType{package}
\name{Bios524hwk6-package}
\alias{Bios524hwk6}
\alias{Bios524hwk6-package}
\title{Bios524hwk6: Homework Functions}
\description{
Includes functions for pythagorean theorem and trimmed mean
}
\keyword{internal}
