% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pythagorian_theorem.R
\name{Pythagorian_theorem}
\alias{Pythagorian_theorem}
\title{Calculate the length of the third side of a right triangle}
\usage{
Pythagorian_theorem(a, b, c)
}
\arguments{
\item{a}{Numeric. Length of one side}

\item{b}{Numeric. Length of the other side}
}
\value{
Numeric. Length of the third side
}
\description{
This function computes the length of the third side in a right triangle
by applying the Pythagorean theorem
}
\examples{
Pythagorian_theorem(3,4,NA)
Pythagorian_theorem(a=4,b=3,c=NA)
}
