% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Calculate the trimmed mean of a numeric vector}
\usage{
trimmed_mean(s, l, y)
}
\arguments{
\item{s}{Number of smallest value to ignore}

\item{l}{Number of largest value to ignore}

\item{y}{Numeric vector}
}
\value{
Numeric.Trimmed mean of the vector
}
\description{
This function computes the average of a numeric vector, where it omits the
's' smallest values and the 'l' largest values from consideration,resulting
what is known as a 'trimmed mean'
}
\examples{
trimmed_mean(10,9,1:30)
}
