% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\docType{data}
\name{y}
\alias{y}
\title{Calculate the trimmed mean of a numeric vector}
\format{
An object of class \code{numeric} of length 50.
}
\usage{
y
}
\arguments{
\item{y}{Numeric vector}

\item{s}{Number of smallest value to ignore}

\item{l}{Number of largest value to ignore}
}
\value{
Numeric.Trimmed mean of the vector
}
\description{
This function computes the average of a numeric vector, where it omits the
's' smallest values and the 'l' largest values from consideration,resulting
what is known as a 'trimmed mean'
}
\examples{
trimmed_mean(10,9,y)
}
\keyword{datasets}
