% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{trimmed_mean}
\usage{
trimmed_mean(x, s, l)
}
\arguments{
\item{x}{numeric vector}

\item{s}{lower trim}

\item{l}{upper trim}
}
\description{
Computes trimmed mean of a numeric vector based on s,l inputs
}
\examples{
trimmed_mean(1:10,1,1)
}
\author{
Samuel Soon
}
